abstract createModule
