+++ web/src/pages/demo/NestedModuleTest/data.d.ts
@@ -1,11 +1,12 @@
 import { Layout } from "react-grid-layout";
+import { Content } from "../../../components/Gallery/GalleryDataType";
 
 export interface tabItem extends Layout {
     text: string,
 }
 export interface Module {
     id: string
-    module: ModulePanelProps
+    module: JSX.Element
 }
 export interface ContentWithId {
     id: string,
